backend/requirements.txt
# China Political Interpretation - Server Dependencies
# =====================================================
# LLM-Powered Political Analysis System

# ===================
# LLM Client (Required)
# ===================
aiohttp>=3.9.0
python-dotenv>=1.0.0

# ===================
# Web Framework
# ===================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# ===================
# Content Extraction
# ===================
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.41.0

# ===================
# OCR (optional)
# ===================
# pytesseract>=0.3.10
# paddleocr>=2.7.0
# paddlepaddle>=2.6.0

# ===================
# PDF Generation
# ===================
reportlab>=4.0.0
# weasyprint>=60.0  # Alternative PDF engine

# ===================
# PDF Reading
# ===================
PyMuPDF>=1.23.0
# pdf2image>=1.17.0  # For PDF OCR fallback

# ===================
# Image Processing
# ===================
Pillow>=10.0.0
